LU-254: don't mark long extent symlinks bad
Long symlinks with the EXT4_EXTENTS_FL set should no longer be considered
as corrupt, since these are created by default with new kernels using ext4
as the filesystem type.  This has not impacted Lustre in the past because
extents are only enabled on the OST, while symlinks are only created on the
MDT where extents are never enabled.

However, now that distros are using ext4 as the base filesystem type, if
the lustre-patched e2fsprogs RPM is installed it will incorrectly consider
all such long symlinks as bad and remove them.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>