LU-367 Handle DB->open errors without crashing
Remove the DB_INIT_LOCK call from the second db->open() call, since
this caused errors on the newer db4.  We shouldn't need this anyway
because the databases are never opened concurrently in r/w mode.

If there is an error opening the MDSDB or OSTDB, print out a consistent
error message.  In lfsck_run_checks() it didn't have enough function
arguments, so when the MDSDB failed to open (e.g. due to a different
version of db4 installed on the MDS, OSS, or client nodes) it caused
lfsck to crash afterward.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I65f9de6886c401315530d6b9e402ce8b1f84e8ce