LU-517 tests: don't fail mkstemp() on Lustre fs
The tst_read_ea test tries to find a filesystem with "user_xattr"
as a mount option (which the MDS mountpoint matches) and then tries
to create a temporary file in /mnt/mds via the VFS and read it back
via libext2fs and verify the value.  Unfortunately, it is not
possible to create files in /mnt/mds, so the test fails.

If mkstemp() fails, just continue looking for a suitable ext3/4
filesystem to run the test on.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I4a11b9edd7adf3cc76f364e3a900811aa242b0c1