LU-752 lfsck: compare EA name correctly
In e2fsck_lfsck_find_ea(), it compares EA name by calling strncmp on the
EA name pairs directly, without length checking. This is wrong because
the passed in EA entry could be an empty entry with 0 EA name length,
or it's name could be a sub-string of the comparing one.

Singed-off-by: Ken Hornstein <kenh@cmf.nrl.navy.mil>
Signed-off-by: Niu Yawei <niu@whamcloud.com>
Change-Id: Ib79ea217e67a6702eee2f41155562a775f6c1578