Set chaos version number.
Put the chaos version before wc's version

Lets us be certain that version numbers are monitonically
increasing (as rpm sees it) without using the Epoch
field in the spec file.  The Epoch field might have caused
issues in the future, which is why I decided to just
reverse our version with Whamcloud's.

Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>